+++ src/base_dumper/common/__init__.py
@@ -13,7 +13,7 @@ from .errors import (
     BaseDumperValueError,
 )
 from .generate_name import random_name
-from .info import debug_info
+from .info import DebugInfo
 from .isolations import IsolationLevel
 from .logger import DumperLogger
 from .mode_level import DumperMode
@@ -28,12 +28,12 @@ __all__ = (
     "BaseDumperValueError",
     "DBConnector",
     "DBMetadata",
+    "DebugInfo",
     "DumperLogger",
     "DumperMode",
     "ExampleReader",
     "IsolationLevel",
     "chunk_query",
-    "debug_info",
     "random_name",
     "timeouts",
     "transfer_diagram",